# openapi/analytics-admin.openapi.json
{
  "components": {
...
    "securitySchemes": {
      "metrics": { "bearerFormat": "JWT", "scheme": "bearer", "type": "http" },
      "session": {
        "in": "cookie",
        "name": "__Host-analytics_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Tenant-safe operational and sales ingestion API",
    "title": "Analytics Admin API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/audit-events": {
      "get": {
        "operationId": "TeamController_auditEvents_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Team"]
      }
    },
...
    "/api/v1/organizations/{organizationId}/team": {
      "get": {
        "operationId": "TeamController_team_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Team"]
      },
      "post": {
        "operationId": "TeamController_addMember_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "tags": ["Team"]
      }
    },
    "/api/v1/organizations/{organizationId}/team/{userId}": {
      "delete": {
        "operationId": "TeamController_removeMember_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "tags": ["Team"]
      },
      "patch": {
        "operationId": "TeamController_updateMember_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "tags": ["Team"]
      }
    },
...
  },
  "servers": [],
...
}